+++ R/package.R
@@ -32,9 +32,9 @@ NULL
 #' for more information.
 #'
 #' @details This package contains several vignettes that are designed to
-#'   showcase its functionality. To view them, type of the command
+#'   showcase its functionality. To view them, please use the code
 #'   `vignette("name", package = "prioritizr")` where `"name"` is the
-#'   name of the desired vignette (e.g. `"gurobi_installation"`.
+#'   name of the desired vignette (e.g. `"gurobi_installation"`).
 #'
 #'   \describe{
 #'