+++ R/solvers.R
@@ -8,7 +8,7 @@ NULL
 #' software currently installed on the system will be used.
 #' For information on the performance of different solvers,
 #' please see Schuster _et al._ (2020) for benchmarks comparing the
-#' run time and solution quality of different solvers when applied to
+#' run time and solution quality of some of these solvers when applied to
 #' different sized datasets.
 #'
 #' @details The following solvers can be used to find solutions for a
@@ -16,14 +16,15 @@ NULL
 #'
 #'   \describe{
 #'
-#'   \item{`add_default_solver`}{This solver uses the best software
+#'   \item{[add_default_solver()]}{This solver uses the best software
 #'     currently installed on the system.}
 #'
 #'   \item{[add_gurobi_solver()]}{[*Gurobi*](https://www.gurobi.com/)
 #'     is a state-of-the-art commercial optimization software with an R package
-#'     interface. It is by far the fastest of the solvers available for
+#'     interface. We recommend using this solver if at all possible.
+#'     It is by far the fastest of the solvers available for
 #'     generating prioritizations, however, it is not freely available. That
-#'    said, licenses are available to academics at no cost. The
+#'     said, licenses are available to academics at no cost. The
 #'     \pkg{gurobi} package is distributed with the *Gurobi* software
 #'     suite. This solver uses the \pkg{gurobi} package to solve problems.}
 #'
@@ -36,22 +37,37 @@ NULL
 #'     This solver uses the \pkg{cplexAPI} package to solve problems using
 #'     *IBM CPLEX*.}
 #'
-#'   \item{[add_rsymphony_solver()]}{
-#'     [*SYMPHONY*](https://projects.coin-or.org/SYMPHONY) is an
-#'     open-source integer programming solver that is part of the Computational
-#'     Infrastructure for Operations Research (COIN-OR) project, an initiative
-#'     to promote development of open-source tools for operations research (a
-#'     field that includes linear programming). The \pkg{Rsymphony} package
-#'     provides an interface to COIN-OR and is available on CRAN. This solver
-#'     uses the \pkg{Rsymphony} package to solve problems.}
+#'  \item{[add_cbc_solver()]}{[*CBC*](https://projects.coin-or.org/Cbc) is an
+#'    open-source mixed integer programming solver that is part of the
+#'     Computational Infrastructure for Operations Research (COIN-OR) project.
+#'     Preliminary benchmarks indicate that it is the fastest open source
+#'     solver currently supported.
+#'     We recommend using this solver if *Gurobi* and *IBM CPLEX* are
+#'     unavailable.
+#'     It requires the \pkg{rcbc} package, which is currently only available on
+#'     [GitHub](https://github.com/dirkschumacher/rcbc).
+#'  }
 #'
-#'   \item{[add_lpsymphony_solver()]}{The \pkg{lpsymphony} package
-#'     provides a different interface to the COIN-OR software suite. Unlike the
-#'     \pkg{Rsymhpony} package, the \pkg{lpsymphony} package is distributed
-#'     through
+#'   \item{[add_lpsymphony_solver()]}{
+#'     [*SYMPHONY*](https://projects.coin-or.org/SYMPHONY) is an
+#'     open-source mixed integer programming solver that is also part of the
+#'     COIN-OR project. Although both *SYMPHONY* and *CBC* are part of
+#'     the COIN-OR project, they are different software.
+#'     The \pkg{lpsymphony} package provides an interface to the *SYMPHONY*
+#'     software, and is distributed through
 #'     [Bioconductor](https://doi.org/doi:10.18129/B9.bioc.lpsymphony).
-#'     The \pkg{lpsymphony} package may be easier to install on Windows or
-#'     Max OSX systems than the \pkg{Rsymphony} package.}
+#'     We recommend using this solver if the CBC solver cannot be installed.
+#'     This solver can use parallel processing to solve problems, so it is
+#'     faster than \pkg{Rsymphony} package interface (see below).
+#'  }
+#'
+#'   \item{[add_rsymphony_solver()]}{
+#'     This solver provides an alternative interface to the
+#'     [*SYMPHONY*](https://projects.coin-or.org/SYMPHONY) solver using
+#'     the \pkg{Rsymphony} package.
+#'     Unlike other solvers, the \pkg{Rsymphony} package can be installed
+#'     directly from the Comprehensive R Archive Network (CRAN).
+#'     It is also the slowest of the available solvers.}
 #'
 #' }
 #'
@@ -76,7 +92,7 @@ NULL
 #' p <- problem(sim_pu_raster, sim_features) %>%
 #'   add_min_set_objective() %>%
 #'   add_relative_targets(0.1) %>%
-#'   add_binary_decisions()
+#'   add_proportion_decisions()
 #'
 #' # create vector to store plot titles
 #' titles <- c()
@@ -84,43 +100,38 @@ NULL
 #' # create empty stack to store solutions
 #' s <- stack()
 #'
-#' # create problem with added rsymphony solver and limit the time spent
-#' # searching for the optimal solution to 2 seconds
-#' if (require("Rsymphony")) {
-#'   titles <- c(titles, "Rsymphony (2s)")
-#'   p1 <- p %>% add_rsymphony_solver(time_limit = 2, verbose = FALSE)
-#'   s <- addLayer(s, solve(p1))
-#' }
-#'
-#' # create problem with added rsymphony solver and limit the time spent
-#' # searching for the optimal solution to 5 seconds
-#' if (require("Rsymphony")) {
-#'   titles <- c(titles, "Rsymphony (5s)")
-#'   p2 <- p %>% add_rsymphony_solver(time_limit = 5, verbose = FALSE)
-#'   s <- addLayer(s, solve(p2))
-#' }
-#'
-#' # if the gurobi is installed: create problem with added gurobi solver
+#' # if gurobi is installed: create problem with added gurobi solver
 #' if (require("gurobi")) {
-#'   titles <- c(titles, "gurobi (5s)")
-#'   p3 <- p %>% add_gurobi_solver(time_limit = 5, verbose = FALSE)
+#'   titles <- c(titles, "gurobi")
+#'   p3 <- p %>% add_gurobi_solver(verbose = FALSE)
 #'   s <- addLayer(s, solve(p3))
 #' }
 #'
-#' # if the cplexAPI is installed: create problem with added cplex solver
+#' # if cplexAPI is installed: create problem with added CPLEX solver
 #' if (require("cplexAPI")) {
-#'   titles <- c(titles, "cplexAPI (5s)")
-#'   p4 <- p %>% add_cplex_solver(time_limit = 5, verbose = FALSE)
+#'   titles <- c(titles, "CPLEX")
+#'   p4 <- p %>% add_cplex_solver(verbose = FALSE)
 #'   s <- addLayer(s, solve(p4))
 #' }
 #'
-#' # if the lpsymphony is installed: create problem with added lpsymphony solver
-#' # note that this solver is skipped on Linux systems due to instability
-#' # issues
-#' if (require("lpsymphony") &
-#'     isTRUE(Sys.info()[["sysname"]] != "Linux")) {
+#' # if rcbc is installed: create problem with added cbc solver
+#' if (require("rcbc")) {
+#'   titles <- c(titles, "CBC")
+#'   p6 <- p %>% add_cbc_solver(verbose = FALSE)
+#'   s <- addLayer(s, solve(p6))
+#' }
+#'
+#' # create problem with added rsymphony solver
+#' if (require("Rsymphony")) {
+#'   titles <- c(titles, "Rsymphony")
+#'   p2 <- p %>% add_rsymphony_solver(verbose = FALSE)
+#'   s <- addLayer(s, solve(p2))
+#' }
+#'
+#' # if lpsymphony is installed: create problem with added lpsymphony solver
+#' if (require("lpsymphony")) {
 #'   titles <- c(titles, "lpsymphony")
-#'   p5 <- p %>% add_lpsymphony_solver(time_limit = 10, verbose = FALSE)
+#'   p5 <- p %>% add_lpsymphony_solver(verbose = FALSE)
 #'   s <- addLayer(s, solve(p5))
 #' }
 #'